Add debounce RC filter to on/off switch
3V3_EN has a 100k pull-up.
RT6150B-33GQW registers 0.4V as logic-low (off).
Debounce resistor will create voltage divider:

  Vout > (Rl * Vcc) / (Rl + Rh)
  0.4V > (Rl * 5V) / (Rl + 100k)
  Rl < 8696

Voltage at divider for Rl of 470ohm:

  Vout = Rl * Vcc / (Rl + Rh)
  Vout = 470 * 5V / (470 + 100k)
  Vout = 0.023V

RT6150B-33GQW registers 1.2V as logic-high (on).
Capacitance for >1ms DISCHARGE (on) debounce time:

  Vc < Vs * (e ^ (-t / (R * C)))
  (5V - 1.2V) < (5V - 0.023V) * (e ^ (-1ms / (100k * C)))
  C > 37.1nF

Selected:

  R: 470
  C: 0.1uF

  (Both are values which already exist in schematic/BOM)

Which results in a discharge (on) time of:

  Vc < Vs * (e ^ (-t / (R * C)))
  (5V - 1.2V) = (5V - 0.023V) * (e ^ (-t / (100k * 0.1uF)))
  t(on) = 2.7ms

Power off (charge) will not be as graceful:

  Vc < Vs * (1 - e ^ (-t / (R * C)))
  (5V - 0.4V) = (5V - 0.023V) * (1 - e ^ (-t / (470 * 0.1uF)))
  t(off) = 121us

But it's probably still sufficient at filtering bounces,
and graceful shutdown is not as important as startup.

<https://datasheets.raspberrypi.com/pico/pico-datasheet.pdf>
<https://www.richtek.com/assets/product_file/RT6150A=RT6150B/DS6150AB-05.pdf>